\name{el_mat}
\alias{el_mat}
\title{Transforms edgelist into matrix}
\usage{
  el_mat(edgelist, type = "weighted")
}
\arguments{
  \item{edgelist}{3-column matrix with node in, node out,
  weight of interaction}

  \item{type}{\code{binary} returns a binary matrix and
  \code{weighted} returns a weighted matrix}
}
\description{
  Transforms edgelist into matrix
}
\author{
  Mauricio Cantor
}

